be/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
python-multipart==0.0.6
SpeechRecognition==3.10.0
aiohttp==3.9.1
python-jose==3.3.0 